refactor: extract claude_session.rs and file_ops.rs from commands (#127 step 9)
- claude_session.rs: 12 functions for Claude Code session detection,
  process tree traversal, and startup command validation
- file_ops.rs: 6 functions for file viewer, directory listing, and
  settings file opening

commands/mod.rs reduced by ~985 lines. 393 tests passing.
