+++ src-tauri/src/commands/mod.rs
@@ -1,4 +1,8 @@
+mod claude_session;
+mod file_ops;
 mod terminal;
+pub use claude_session::*;
+pub use file_ops::*;
 pub use terminal::*;
 
 use std::collections::HashMap;
@@ -21,372 +25,6 @@ use crate::terminal::{
 #[tauri::command]
 pub fn greet(name: &str) -> String {
     format!("Hello, {}! Welcome to Laymux.", name)
-}
-
-/// Resolve Claude Code session IDs for known Claude terminals.
-///
-/// Returns a map of terminal ID → Claude session ID by:
-/// 1. (Primary) PID tree matching: walk the process tree from the PTY child PID
-///    and match against `~/.claude/sessions/<pid>.json` files.
-/// 2. (Fallback) CWD + most-recent matching: compare the terminal's CWD with
-///    session file CWD, picking the most recently started session.
-#[tauri::command]
-pub fn get_claude_session_ids(
-    session_max_age_hours: Option<u64>,
-    state: State<Arc<AppState>>,
-) -> Result<HashMap<String, String>, String> {
-    let known: Vec<String> = {
-        let k = state.known_claude_terminals.lock_or_err()?;
-        k.iter().cloned().collect()
-    };
-
-    if known.is_empty() {
-        return Ok(HashMap::new());
-    }
-
-    // Read session files from ~/.claude/sessions/
-    let sessions_dir = resolve_claude_sessions_dir();
-    let session_files = read_claude_session_files(&sessions_dir, session_max_age_hours);
-
-    let mut result = HashMap::new();
-
-    for terminal_id in &known {
-        // Get child PID from PTY handle
-        let child_pid = {
-            let ptys = state.pty_handles.lock_or_err()?;
-            ptys.get(terminal_id).and_then(|h| h.child_pid())
-        };
-
-        // Get terminal CWD for fallback
-        let terminal_cwd = {
-            let terminals = state.terminals.lock_or_err()?;
-            terminals.get(terminal_id).and_then(|s| s.cwd.clone())
-        };
-
-        // Strategy 1: PID tree matching
-        if let Some(pid) = child_pid {
-            let descendant_pids = get_descendant_pids(pid);
-            if let Some(session_id) = find_session_by_pids(&session_files, &descendant_pids) {
-                result.insert(terminal_id.clone(), session_id);
-                continue;
-            }
-        }
-
-        // Strategy 2: CWD + most-recent fallback
-        if let Some(ref cwd) = terminal_cwd {
-            if let Some(session_id) = find_session_by_cwd(&session_files, cwd) {
-                eprintln!(
-                    "[claude-session] PID tree match failed for {terminal_id}, \
-                     using CWD fallback (cwd={cwd})"
-                );
-                result.insert(terminal_id.clone(), session_id);
-            }
-        }
-    }
-
-    Ok(result)
-}
-
-/// A parsed Claude session file entry.
-#[derive(Debug, Clone)]
-struct ClaudeSessionFile {
-    pid: u32,
-    session_id: String,
-    cwd: String,
-    started_at: u64,
-}
-
-/// Validate that a startup command override is safe to execute.
-///
-/// Allowed forms:
-/// - `claude --resume <valid_session_id>` — Claude session restore
-/// - `<viewer_command> '<file_path>'` — Extension viewer from settings whitelist
-///
-/// `allowed_viewer_commands` should contain the command names registered in
-/// `settings.fileExplorer.extensionViewers` (e.g., `["vi", "less"]`).
-pub(crate) fn is_valid_startup_command_override(
-    cmd: &str,
-    allowed_viewer_commands: &[String],
-) -> bool {
-    // Check claude --resume pattern first
-    if cmd
-        .strip_prefix("claude --resume ")
-        .is_some_and(|id| is_valid_session_id(id))
-    {
-        return true;
-    }
-
-    // Check extension viewer pattern: "<command> '<path>'"
-    // The path is single-quoted by shellEscape on the frontend.
-    // shellEscape escapes embedded single quotes as '\'' (end-quote, escaped-quote, start-quote),
-    // so the full argument looks like: 'part1'\''part2'
-    // We validate the structure: command + space + shell-escaped path, no other shell metacharacters.
-    for viewer_cmd in allowed_viewer_commands {
-        if let Some(rest) = cmd.strip_prefix(viewer_cmd.as_str()) {
-            if let Some(after_space) = rest.strip_prefix(' ') {
-                if is_valid_shell_escaped_path(after_space) {
-                    return true;
-                }
-            }
-        }
-    }
-
-    false
-}
-
-/// Validate a shell-escaped path produced by the frontend's `shellEscape()`.
-///
-/// Accepted format: `'<content>'` where embedded single-quotes are escaped as `'\''`.
-/// The full pattern is one or more `'...'` segments separated by `'\''`.
-/// No shell metacharacters (`;`, `&`, `|`, `$`, `` ` ``, `(`, `)`) are allowed
-/// inside the quoted segments.
-fn is_valid_shell_escaped_path(s: &str) -> bool {
-    // Must start with ' and end with '
-    if !s.starts_with('\'') || !s.ends_with('\'') || s.len() < 2 {
-        return false;
-    }
-
-    // Dangerous shell metacharacters that must not appear even inside quotes.
-    // Single-quoted strings in POSIX shells don't interpret these, but since
-    // the escaped-quote pattern ('\'') temporarily leaves the quoted context,
-    // we reject them to be safe.
-    const DANGEROUS: &[char] = &[';', '&', '|', '$', '`', '(', ')', '\n'];
-
-    // Strip outer quotes and check segments split by '\'' (escaped quote)
-    let inner = &s[1..s.len() - 1];
-    for segment in inner.split("'\\''") {
-        if segment.contains(DANGEROUS) {
-            return false;
-        }
-    }
-
-    true
-}
-
-/// Validate that a Claude session ID contains only safe characters
-/// (alphanumeric, hyphens, underscores). Prevents command injection when
-/// the ID is interpolated into `claude --resume <id>`.
-fn is_valid_session_id(id: &str) -> bool {
-    !id.is_empty()
-        && id
-            .chars()
-            .all(|c| c.is_ascii_alphanumeric() || c == '-' || c == '_')
-}
-
-/// Resolve the Claude sessions directory path.
-fn resolve_claude_sessions_dir() -> std::path::PathBuf {
-    #[cfg(windows)]
-    {
-        if let Ok(home) = std::env::var("USERPROFILE") {
-            return std::path::PathBuf::from(home)
-                .join(".claude")
-                .join("sessions");
-        }
-    }
-    #[cfg(not(windows))]
-    {
-        if let Ok(home) = std::env::var("HOME") {
-            return std::path::PathBuf::from(home)
-                .join(".claude")
-                .join("sessions");
-        }
-    }
-    std::path::PathBuf::from(".claude").join("sessions")
-}
-
-/// Read and parse all Claude session files from the given directory.
-/// If `max_age_hours` is Some, sessions older than the threshold are filtered out.
-fn read_claude_session_files(
-    dir: &std::path::Path,
-    max_age_hours: Option<u64>,
-) -> Vec<ClaudeSessionFile> {
-    let entries = match std::fs::read_dir(dir) {
-        Ok(e) => e,
-        Err(_) => return Vec::new(),
-    };
-
-    // Compute the cutoff timestamp (seconds since epoch) if max_age_hours is set.
-    // 0 means "no filter" (accept all sessions regardless of age).
-    let cutoff = max_age_hours.filter(|&hours| hours > 0).and_then(|hours| {
-        std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .ok()
-            .map(|d| d.as_secs().saturating_sub(hours * 3600))
-    });
-
-    let mut result = Vec::new();
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) != Some("json") {
-            continue;
-        }
-        if let Ok(content) = std::fs::read_to_string(&path) {
-            if let Ok(val) = serde_json::from_str::<serde_json::Value>(&content) {
-                let pid = val.get("pid").and_then(|v| v.as_u64()).unwrap_or(0) as u32;
-                let session_id = val
-                    .get("sessionId")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let cwd = val
-                    .get("cwd")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
-                let started_at = val.get("startedAt").and_then(|v| v.as_u64()).unwrap_or(0);
-
-                // Skip stale sessions
-                if let Some(min_ts) = cutoff {
-                    if started_at < min_ts {
-                        continue;
-                    }
-                }
-
-                if is_valid_session_id(&session_id) {
-                    result.push(ClaudeSessionFile {
-                        pid,
-                        session_id,
-                        cwd,
-                        started_at,
-                    });
-                }
-            }
-        }
-    }
-    result
-}
-
-/// Get all descendant PIDs of a given process (including the process itself).
-/// Uses platform-specific process enumeration.
-fn get_descendant_pids(root_pid: u32) -> Vec<u32> {
-    use std::collections::{HashSet, VecDeque};
-
-    let mut result = vec![root_pid];
-    let mut seen = HashSet::new();
-    seen.insert(root_pid);
-
-    #[cfg(windows)]
-    {
-        match create_process_snapshot() {
-            Ok(snapshot) => {
-                let parent_map = build_parent_map(&snapshot);
-                let mut queue = VecDeque::new();
-                queue.push_back(root_pid);
-                while let Some(pid) = queue.pop_front() {
-                    if let Some(children) = parent_map.get(&pid) {
-                        for &child in children {
-                            if seen.insert(child) {
-                                result.push(child);
-                                queue.push_back(child);
-                            }
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                eprintln!(
-                    "[claude-session] Failed to create process snapshot for PID {root_pid}: {e}"
-                );
-            }
-        }
-    }
-
-    #[cfg(not(windows))]
-    {
-        let mut queue = VecDeque::new();
-        queue.push_back(root_pid);
-        while let Some(pid) = queue.pop_front() {
-            let children_path = format!("/proc/{pid}/task/{pid}/children");
-            match std::fs::read_to_string(&children_path) {
-                Ok(content) => {
-                    for token in content.split_whitespace() {
-                        if let Ok(child_pid) = token.parse::<u32>() {
-                            if seen.insert(child_pid) {
-                                result.push(child_pid);
-                                queue.push_back(child_pid);
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    eprintln!(
-                        "[claude-session] Cannot read /proc children for PID {pid}: {e} \
-                         (kernel CONFIG_PROC_CHILDREN may be disabled, or PID exited)"
-                    );
-                }
-            }
-        }
-    }
-
-    result
-}
-
-#[cfg(windows)]
-fn create_process_snapshot() -> Result<Vec<(u32, u32)>, String> {
-    use windows_sys::Win32::Foundation::{CloseHandle, INVALID_HANDLE_VALUE};
-    use windows_sys::Win32::System::Diagnostics::ToolHelp::{
-        CreateToolhelp32Snapshot, Process32First, Process32Next, PROCESSENTRY32, TH32CS_SNAPPROCESS,
-    };
-
-    /// RAII guard that closes a Windows HANDLE on drop, preventing leaks on panic.
-    struct SnapshotGuard(windows_sys::Win32::Foundation::HANDLE);
-    impl Drop for SnapshotGuard {
-        fn drop(&mut self) {
-            unsafe {
-                CloseHandle(self.0);
-            }
-        }
-    }
-
-    unsafe {
-        let snap = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
-        if snap == INVALID_HANDLE_VALUE {
-            return Err("Failed to create snapshot".into());
-        }
-        let _guard = SnapshotGuard(snap);
-
-        let mut entry: PROCESSENTRY32 = std::mem::zeroed();
-        entry.dwSize = std::mem::size_of::<PROCESSENTRY32>() as u32;
-        let mut pairs = Vec::new();
-        if Process32First(snap, &mut entry) != 0 {
-            loop {
-                pairs.push((entry.th32ProcessID, entry.th32ParentProcessID));
-                if Process32Next(snap, &mut entry) == 0 {
-                    break;
-                }
-            }
-        }
-        Ok(pairs)
-    }
-}
-
-#[cfg(windows)]
-fn build_parent_map(snapshot: &[(u32, u32)]) -> HashMap<u32, Vec<u32>> {
-    let mut map: HashMap<u32, Vec<u32>> = HashMap::new();
-    for &(pid, ppid) in snapshot {
-        map.entry(ppid).or_default().push(pid);
-    }
-    map
-}
-
-/// Find a Claude session ID by matching any of the given PIDs against session file PIDs.
-/// When multiple sessions match, the most recently started one wins.
-fn find_session_by_pids(sessions: &[ClaudeSessionFile], pids: &[u32]) -> Option<String> {
-    sessions
-        .iter()
-        .filter(|s| pids.contains(&s.pid))
-        .max_by_key(|s| s.started_at)
-        .map(|s| s.session_id.clone())
-}
-
-/// Find a Claude session ID by matching CWD (most recent session wins).
-fn find_session_by_cwd(sessions: &[ClaudeSessionFile], cwd: &str) -> Option<String> {
-    let normalized_cwd = path_utils::normalize_path_for_comparison(cwd);
-    sessions
-        .iter()
-        .filter(|s| path_utils::normalize_path_for_comparison(&s.cwd) == normalized_cwd)
-        .max_by_key(|s| s.started_at)
-        .map(|s| s.session_id.clone())
 }
 
 #[tauri::command]
@@ -930,268 +568,6 @@ pub fn load_memo(key: String) -> Result<String, String> {
 #[tauri::command]
 pub fn save_memo(key: String, content: String) -> Result<(), String> {
     crate::settings::save_memo(&key, &content)
-}
-
-#[tauri::command]
-pub fn open_settings_file() -> Result<(), String> {
-    let path = crate::settings::settings_path();
-    #[cfg(target_os = "windows")]
-    {
-        crate::process::headless_command("cmd")
-            .args(["/C", "start", "", &path.to_string_lossy()])
-            .spawn()
-            .map_err(|e| format!("Failed to open settings.json: {e}"))?;
-    }
-    #[cfg(target_os = "linux")]
-    {
-        std::process::Command::new("xdg-open")
-            .arg(&path)
-            .spawn()
-            .map_err(|e| format!("Failed to open settings.json: {e}"))?;
-    }
-    Ok(())
-}
-
-/// Content type classification for file viewer.
-#[derive(Debug, Clone, serde::Serialize)]
-#[serde(rename_all = "camelCase", tag = "kind")]
-pub enum FileViewerContent {
-    /// Text file — content included inline.
-    Text { content: String, truncated: bool },
-    /// Image file — inline data URL (base64).
-    Image {
-        #[serde(rename = "dataUrl")]
-        data_url: String,
-    },
-    /// Binary/unsupported — show info only.
-    Binary { size: u64 },
-}
-
-const IMAGE_EXTENSIONS: &[&str] = &[
-    ".png", ".jpg", ".jpeg", ".gif", ".svg", ".webp", ".bmp", ".ico",
-];
-
-const TEXT_EXTENSIONS: &[&str] = &[
-    ".txt",
-    ".md",
-    ".json",
-    ".jsonc",
-    ".js",
-    ".jsx",
-    ".ts",
-    ".tsx",
-    ".rs",
-    ".py",
-    ".go",
-    ".c",
-    ".cpp",
-    ".h",
-    ".hpp",
-    ".toml",
-    ".yaml",
-    ".yml",
-    ".xml",
-    ".html",
-    ".css",
-    ".scss",
-    ".sh",
-    ".bash",
-    ".zsh",
-    ".fish",
-    ".bat",
-    ".ps1",
-    ".log",
-    ".env",
-    ".gitignore",
-    ".editorconfig",
-    ".conf",
-    ".cfg",
-    ".ini",
-    ".csv",
-];
-
-/// Read a file and classify it for the file viewer.
-#[tauri::command]
-pub fn read_file_for_viewer(
-    path: String,
-    max_bytes: Option<usize>,
-) -> Result<FileViewerContent, String> {
-    // Resolve WSL paths on Windows
-    let distro = if cfg!(windows) && path.starts_with('/') && !path.starts_with("/mnt/") {
-        path_utils::get_default_wsl_distro()
-    } else {
-        None
-    };
-    let resolved = path_utils::resolve_path_for_windows(&path, distro.as_deref());
-    let file_path = std::path::Path::new(&resolved);
-    let ext = file_path
-        .extension()
-        .and_then(|e| e.to_str())
-        .map(|e| format!(".{}", e.to_lowercase()))
-        .unwrap_or_default();
-
-    if IMAGE_EXTENSIONS.contains(&ext.as_str()) {
-        // Read image and return as data URL (convertFileSrc can't handle WSL UNC paths)
-        let bytes = std::fs::read(&resolved).map_err(|e| format!("Cannot read image: {e}"))?;
-        let mime = match ext.as_str() {
-            ".png" => "image/png",
-            ".jpg" | ".jpeg" => "image/jpeg",
-            ".gif" => "image/gif",
-            ".svg" => "image/svg+xml",
-            ".webp" => "image/webp",
-            ".bmp" => "image/bmp",
-            ".ico" => "image/x-icon",
-            _ => "application/octet-stream",
-        };
-        let b64 = base64_encode(&bytes);
-        return Ok(FileViewerContent::Image {
-            data_url: format!("data:{mime};base64,{b64}"),
-        });
-    }
-
-    let metadata = std::fs::metadata(&resolved).map_err(|e| format!("Cannot stat file: {e}"))?;
-    let size = metadata.len();
-    let limit = max_bytes.unwrap_or(1_048_576) as u64; // 1MB default
-
-    // Treat known text extensions or small files as text
-    let is_text_ext = TEXT_EXTENSIONS.contains(&ext.as_str()) || ext.is_empty();
-    if !is_text_ext && size > limit {
-        return Ok(FileViewerContent::Binary { size });
-    }
-
-    // Read only up to limit bytes (avoid loading entire large files into memory)
-    let read_limit = std::cmp::min(size, limit) as usize;
-    let truncated = size > limit;
-    let mut buf = vec![0u8; read_limit];
-    {
-        use std::io::Read;
-        let mut f = std::fs::File::open(&resolved).map_err(|e| format!("Cannot open file: {e}"))?;
-        f.read_exact(&mut buf)
-            .map_err(|e| format!("Cannot read file: {e}"))?;
-    }
-
-    match std::str::from_utf8(&buf) {
-        Ok(text) => Ok(FileViewerContent::Text {
-            content: text.to_string(),
-            truncated,
-        }),
-        Err(_) if is_text_ext => {
-            // Lossy conversion for known text extensions
-            Ok(FileViewerContent::Text {
-                content: String::from_utf8_lossy(&buf).into_owned(),
-                truncated,
-            })
-        }
-        Err(_) => Ok(FileViewerContent::Binary { size }),
-    }
-}
-
-/// A single directory entry returned by `list_directory`.
-#[derive(Debug, Clone, serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DirEntry {
-    pub name: String,
-    pub is_directory: bool,
-    pub is_symlink: bool,
-    pub is_executable: bool,
-    pub size: u64,
-}
-
-/// List directory contents and return structured metadata for each entry.
-#[tauri::command]
-pub fn list_directory(path: String, wsl_distro: Option<String>) -> Result<Vec<DirEntry>, String> {
-    // On Windows, resolve Linux paths to UNC paths
-    let distro = wsl_distro.or_else(|| {
-        // Auto-detect WSL distro if path looks like a Linux path
-        if cfg!(windows) && path.starts_with('/') && !path.starts_with("/mnt/") {
-            path_utils::get_default_wsl_distro()
-        } else {
-            None
-        }
-    });
-    let resolved = path_utils::resolve_path_for_windows(&path, distro.as_deref());
-    let dir_path = std::path::Path::new(&resolved);
-    let entries = std::fs::read_dir(dir_path).map_err(|e| format!("Cannot read directory: {e}"))?;
-
-    let mut result = Vec::new();
-    for entry in entries {
-        let entry = match entry {
-            Ok(e) => e,
-            Err(_) => continue, // skip unreadable entries
-        };
-        let name = entry.file_name().to_string_lossy().into_owned();
-
-        // Use symlink_metadata to detect symlinks (metadata follows symlinks)
-        let sym_meta = entry.path().symlink_metadata();
-        let is_symlink = sym_meta.as_ref().map(|m| m.is_symlink()).unwrap_or(false);
-
-        // Follow symlinks for the actual file type and size
-        let meta = entry.path().metadata();
-        let is_directory = meta.as_ref().map(|m| m.is_dir()).unwrap_or(false);
-        let size = meta.as_ref().map(|m| m.len()).unwrap_or(0);
-
-        // Check executable bit (Unix only)
-        #[cfg(unix)]
-        let is_executable = {
-            use std::os::unix::fs::PermissionsExt;
-            meta.as_ref()
-                .map(|m| !m.is_dir() && (m.permissions().mode() & 0o111) != 0)
-                .unwrap_or(false)
-        };
-        #[cfg(not(unix))]
-        let is_executable = {
-            // On Windows, check common executable extensions
-            let ext = entry
-                .path()
-                .extension()
-                .and_then(|e| e.to_str())
-                .map(|e| e.to_lowercase())
-                .unwrap_or_default();
-            !is_directory && matches!(ext.as_str(), "exe" | "cmd" | "bat" | "ps1" | "com")
-        };
-
-        result.push(DirEntry {
-            name,
-            is_directory,
-            is_symlink,
-            is_executable,
-            size,
-        });
-    }
-
-    // Sort: directories first, then alphabetically (case-insensitive)
-    result.sort_by(|a, b| {
-        b.is_directory
-            .cmp(&a.is_directory)
-            .then_with(|| a.name.to_lowercase().cmp(&b.name.to_lowercase()))
-    });
-
-    Ok(result)
-}
-
-/// Simple base64 encoder (no external crate needed).
-fn base64_encode(input: &[u8]) -> String {
-    const TABLE: &[u8; 64] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    let mut result = String::with_capacity(input.len().div_ceil(3) * 4);
-    for chunk in input.chunks(3) {
-        let b0 = chunk[0] as u32;
-        let b1 = if chunk.len() > 1 { chunk[1] as u32 } else { 0 };
-        let b2 = if chunk.len() > 2 { chunk[2] as u32 } else { 0 };
-        let combined = (b0 << 16) | (b1 << 8) | b2;
-        result.push(TABLE[((combined >> 18) & 0x3F) as usize] as char);
-        result.push(TABLE[((combined >> 12) & 0x3F) as usize] as char);
-        if chunk.len() > 1 {
-            result.push(TABLE[((combined >> 6) & 0x3F) as usize] as char);
-        } else {
-            result.push('=');
-        }
-        if chunk.len() > 2 {
-            result.push(TABLE[(combined & 0x3F) as usize] as char);
-        } else {
-            result.push('=');
-        }
-    }
-    result
 }
 
 /// Split an `issueReporter.shell` prefix into tokens, respecting single and double quotes.
@@ -3224,24 +2600,6 @@ mod tests {
     }
 
     #[test]
-    fn base64_encode_empty() {
-        assert_eq!(base64_encode(b""), "");
-    }
-
-    #[test]
-    fn base64_encode_hello() {
-        assert_eq!(base64_encode(b"Hello"), "SGVsbG8=");
-    }
-
-    #[test]
-    fn base64_encode_roundtrip() {
-        let original = b"screenshot data \x00\xff\x80";
-        let encoded = base64_encode(original);
-        let decoded = crate::automation_server::base64_decode(&encoded).unwrap();
-        assert_eq!(decoded, original);
-    }
-
-    #[test]
     fn is_monospace_detects_consolas() {
         use font_kit::source::SystemSource;
         let source = SystemSource::new();
@@ -4086,93 +3444,6 @@ mod tests {
         );
     }
 
-    // -- Claude session file parsing tests --
-
-    #[test]
-    fn read_claude_session_files_empty_dir() {
-        let tmp = tempfile::tempdir().unwrap();
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert!(sessions.is_empty());
-    }
-
-    #[test]
-    fn read_claude_session_files_valid_json() {
-        let tmp = tempfile::tempdir().unwrap();
-        let content = r#"{"pid":12345,"sessionId":"abc-123","cwd":"/home/user","startedAt":1000}"#;
-        std::fs::write(tmp.path().join("12345.json"), content).unwrap();
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert_eq!(sessions.len(), 1);
-        assert_eq!(sessions[0].pid, 12345);
-        assert_eq!(sessions[0].session_id, "abc-123");
-        assert_eq!(sessions[0].cwd, "/home/user");
-        assert_eq!(sessions[0].started_at, 1000);
-    }
-
-    #[test]
-    fn read_claude_session_files_ignores_non_json() {
-        let tmp = tempfile::tempdir().unwrap();
-        std::fs::write(tmp.path().join("readme.txt"), "not json").unwrap();
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert!(sessions.is_empty());
-    }
-
-    #[test]
-    fn read_claude_session_files_ignores_invalid_json() {
-        let tmp = tempfile::tempdir().unwrap();
-        std::fs::write(tmp.path().join("bad.json"), "not valid json!").unwrap();
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert!(sessions.is_empty());
-    }
-
-    #[test]
-    fn find_session_by_pids_matches() {
-        let sessions = vec![
-            ClaudeSessionFile {
-                pid: 100,
-                session_id: "s1".into(),
-                cwd: "/a".into(),
-                started_at: 1,
-            },
-            ClaudeSessionFile {
-                pid: 200,
-                session_id: "s2".into(),
-                cwd: "/b".into(),
-                started_at: 2,
-            },
-        ];
-        assert_eq!(find_session_by_pids(&sessions, &[200]), Some("s2".into()));
-        assert_eq!(find_session_by_pids(&sessions, &[300]), None);
-    }
-
-    #[test]
-    fn find_session_by_cwd_picks_most_recent() {
-        let sessions = vec![
-            ClaudeSessionFile {
-                pid: 1,
-                session_id: "old".into(),
-                cwd: "/home/user".into(),
-                started_at: 100,
-            },
-            ClaudeSessionFile {
-                pid: 2,
-                session_id: "new".into(),
-                cwd: "/home/user".into(),
-                started_at: 200,
-            },
-            ClaudeSessionFile {
-                pid: 3,
-                session_id: "other".into(),
-                cwd: "/other".into(),
-                started_at: 300,
-            },
-        ];
-        assert_eq!(
-            find_session_by_cwd(&sessions, "/home/user"),
-            Some("new".into())
-        );
-        assert_eq!(find_session_by_cwd(&sessions, "/nonexistent"), None);
-    }
-
     #[test]
     fn normalize_path_for_comparison_trims_slashes() {
         assert_eq!(
@@ -4199,262 +3470,6 @@ mod tests {
             path_utils::normalize_path_for_comparison("/mnt/c"),
             path_utils::normalize_path_for_comparison("C:\\")
         );
-    }
-
-    #[test]
-    fn get_descendant_pids_includes_root() {
-        let pids = get_descendant_pids(99999);
-        assert!(pids.contains(&99999));
-    }
-
-    // -- Session ID validation tests --
-
-    #[test]
-    fn is_valid_session_id_accepts_safe_ids() {
-        assert!(is_valid_session_id("abc-123"));
-        assert!(is_valid_session_id("session_id_v2"));
-        assert!(is_valid_session_id("a1b2c3"));
-        assert!(is_valid_session_id("ABC-def_012"));
-    }
-
-    #[test]
-    fn is_valid_session_id_rejects_dangerous_ids() {
-        assert!(!is_valid_session_id(""));
-        assert!(!is_valid_session_id("id; rm -rf /"));
-        assert!(!is_valid_session_id("id && echo pwned"));
-        assert!(!is_valid_session_id("id | cat /etc/passwd"));
-        assert!(!is_valid_session_id("$(whoami)"));
-        assert!(!is_valid_session_id("id`whoami`"));
-        assert!(!is_valid_session_id("hello world"));
-        assert!(!is_valid_session_id("id\nnewline"));
-    }
-
-    // -- Startup command override validation tests --
-
-    #[test]
-    fn startup_command_override_accepts_valid_resume() {
-        let no_viewers: &[String] = &[];
-        assert!(is_valid_startup_command_override(
-            "claude --resume abc-123",
-            no_viewers
-        ));
-        assert!(is_valid_startup_command_override(
-            "claude --resume session_v2",
-            no_viewers,
-        ));
-        assert!(is_valid_startup_command_override(
-            "claude --resume A1B2",
-            no_viewers
-        ));
-    }
-
-    #[test]
-    fn startup_command_override_rejects_arbitrary_commands() {
-        let no_viewers: &[String] = &[];
-        assert!(!is_valid_startup_command_override("rm -rf /", no_viewers));
-        assert!(!is_valid_startup_command_override("echo pwned", no_viewers));
-        assert!(!is_valid_startup_command_override(
-            "claude --resume bad; rm -rf /",
-            no_viewers,
-        ));
-        assert!(!is_valid_startup_command_override(
-            "claude --resume $(whoami)",
-            no_viewers,
-        ));
-        assert!(!is_valid_startup_command_override(
-            "claude --resume id && echo x",
-            no_viewers,
-        ));
-        assert!(!is_valid_startup_command_override("", no_viewers));
-        assert!(!is_valid_startup_command_override(
-            "claude --resume ",
-            no_viewers
-        ));
-        assert!(!is_valid_startup_command_override(
-            "claude --resume",
-            no_viewers
-        ));
-        assert!(!is_valid_startup_command_override(
-            "not-claude --resume abc",
-            no_viewers,
-        ));
-    }
-
-    #[test]
-    fn startup_command_override_accepts_whitelisted_viewer_command() {
-        let viewers = vec!["vi".to_string(), "less".to_string(), "cat".to_string()];
-        // Simple file path
-        assert!(is_valid_startup_command_override(
-            "vi '/home/user/file.txt'",
-            &viewers
-        ));
-        assert!(is_valid_startup_command_override(
-            "less '/tmp/log.log'",
-            &viewers
-        ));
-        assert!(is_valid_startup_command_override(
-            "cat '/data/notes.md'",
-            &viewers
-        ));
-        // Windows-style path
-        assert!(is_valid_startup_command_override(
-            "vi 'C:\\Users\\test\\file.rs'",
-            &viewers
-        ));
-        // Path with spaces
-        assert!(is_valid_startup_command_override(
-            "vi '/home/user/my file.txt'",
-            &viewers
-        ));
-        // Path with embedded single quote (shellEscape produces: 'it'\''s here')
-        assert!(is_valid_startup_command_override(
-            "vi 'it'\\''s here'",
-            &viewers
-        ));
-    }
-
-    #[test]
-    fn startup_command_override_rejects_non_whitelisted_viewer_command() {
-        let viewers = vec!["vi".to_string(), "less".to_string()];
-        // Command not in whitelist
-        assert!(!is_valid_startup_command_override(
-            "rm '/home/user/file.txt'",
-            &viewers
-        ));
-        assert!(!is_valid_startup_command_override(
-            "bash '/tmp/evil.sh'",
-            &viewers
-        ));
-    }
-
-    #[test]
-    fn startup_command_override_rejects_viewer_with_injection() {
-        let viewers = vec!["vi".to_string()];
-        // Injection attempts inside the "path" argument
-        assert!(!is_valid_startup_command_override(
-            "vi 'file.txt'; rm -rf /",
-            &viewers
-        ));
-        assert!(!is_valid_startup_command_override(
-            "vi 'file.txt' && echo pwned",
-            &viewers
-        ));
-        assert!(!is_valid_startup_command_override("vi $(whoami)", &viewers));
-        assert!(!is_valid_startup_command_override(
-            "vi file.txt; rm -rf /",
-            &viewers
-        ));
-        // No argument at all
-        assert!(!is_valid_startup_command_override("vi", &viewers));
-        assert!(!is_valid_startup_command_override("vi ", &viewers));
-    }
-
-    #[test]
-    fn read_claude_session_files_rejects_invalid_session_id() {
-        let tmp = tempfile::tempdir().unwrap();
-        let content = r#"{"pid":1,"sessionId":"bad; rm -rf /","cwd":"/home","startedAt":1}"#;
-        std::fs::write(tmp.path().join("1.json"), content).unwrap();
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert!(sessions.is_empty());
-    }
-
-    #[test]
-    fn find_session_by_pids_picks_most_recent_on_multiple_matches() {
-        let sessions = vec![
-            ClaudeSessionFile {
-                pid: 100,
-                session_id: "old-session".into(),
-                cwd: "/a".into(),
-                started_at: 1,
-            },
-            ClaudeSessionFile {
-                pid: 200,
-                session_id: "new-session".into(),
-                cwd: "/b".into(),
-                started_at: 10,
-            },
-        ];
-        // Both PIDs match — should pick the most recent (started_at=10)
-        assert_eq!(
-            find_session_by_pids(&sessions, &[100, 200]),
-            Some("new-session".into())
-        );
-    }
-
-    // -- Stale session filtering tests --
-
-    #[test]
-    fn read_claude_session_files_filters_stale_sessions() {
-        let tmp = tempfile::tempdir().unwrap();
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        // Recent session (1 hour ago)
-        let recent = format!(
-            r#"{{"pid":1,"sessionId":"recent","cwd":"/a","startedAt":{}}}"#,
-            now - 3600
-        );
-        std::fs::write(tmp.path().join("1.json"), recent).unwrap();
-
-        // Stale session (48 hours ago)
-        let stale = format!(
-            r#"{{"pid":2,"sessionId":"stale","cwd":"/b","startedAt":{}}}"#,
-            now - 48 * 3600
-        );
-        std::fs::write(tmp.path().join("2.json"), stale).unwrap();
-
-        // With 24h max age, only the recent session should pass
-        let sessions = read_claude_session_files(tmp.path(), Some(24));
-        assert_eq!(sessions.len(), 1);
-        assert_eq!(sessions[0].session_id, "recent");
-    }
-
-    #[test]
-    fn read_claude_session_files_no_filter_when_none() {
-        let tmp = tempfile::tempdir().unwrap();
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        // Old session (72 hours ago)
-        let old = format!(
-            r#"{{"pid":1,"sessionId":"old","cwd":"/a","startedAt":{}}}"#,
-            now - 72 * 3600
-        );
-        std::fs::write(tmp.path().join("1.json"), old).unwrap();
-
-        // No max age filter — session should be included
-        let sessions = read_claude_session_files(tmp.path(), None);
-        assert_eq!(sessions.len(), 1);
-        assert_eq!(sessions[0].session_id, "old");
-    }
-
-    #[test]
-    fn read_claude_session_files_zero_hours_disables_filter() {
-        let tmp = tempfile::tempdir().unwrap();
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        // Very old session (30 days ago)
-        let old = format!(
-            r#"{{"pid":1,"sessionId":"ancient","cwd":"/a","startedAt":{}}}"#,
-            now - 30 * 24 * 3600
-        );
-        std::fs::write(tmp.path().join("1.json"), old).unwrap();
-
-        // 0 hours = disabled, but saturating_sub means cutoff = now,
-        // so we actually need to handle 0 as a special case.
-        // Let's verify current behavior: 0 * 3600 = 0, cutoff = now - 0 = now.
-        // startedAt < now → filtered out. That's NOT what we want.
-        // We should treat 0 as "no filter".
-        let sessions = read_claude_session_files(tmp.path(), Some(0));
-        assert_eq!(sessions.len(), 1);
-        assert_eq!(sessions[0].session_id, "ancient");
     }
 
     #[test]
